Revert "chore(flashcard): code refactoring"
This reverts commit 637a75e513cb1c9483d5ba729f8d48948e9e0860.

+++ src/Siklid/Application/Flashcard/DeleteFlashcard.php
@@ -6,7 +6,6 @@ namespace App\Siklid\Application\Flashcard;
 
 use App\Foundation\Action\AbstractAction;
 use App\Foundation\Http\Request;
-use App\Siklid\Application\Contract\Entity\FlashCardInterface;
 use App\Siklid\Document\Flashcard;
 use App\Siklid\Security\UserResolverInterface as UserResolver;
 use Doctrine\ODM\MongoDB\DocumentManager as DM;
@@ -26,14 +25,18 @@ class DeleteFlashcard extends AbstractAction
         $this->userResolver = $userResolver;
     }
 
-    public function execute(): FlashCardInterface
+    public function execute(): mixed
     {
         $flashcardId = (string)$this->request->get('id');
         $flashcard = $this->dm->getRepository(Flashcard::class)->find($flashcardId);
 
-        assert($flashcard instanceof Flashcard, new NotFoundHttpException('Flashcard not found'));
+        if (! $flashcard instanceof Flashcard) {
+            throw new NotFoundHttpException('Flashcard not found');
+        }
 
-        assert($flashcard->getUser() === $this->userResolver->getUser(), new UnauthorizedHttpException('You are not allowed to delete this flashcard'));
+        if ($flashcard->getUser() !== $this->userResolver->getUser()) {
+            throw new UnauthorizedHttpException('You are not allowed to delete this flashcard');
+        }
 
         $flashcard->delete();
         $this->dm->persist($flashcard);
